+++ core/src/test/java/io/grpc/inprocess/InProcessClientTransportFactoryTest.java
@@ -16,8 +16,8 @@
 
 package io.grpc.inprocess;
 
+import io.grpc.internal.AbstractClientTransportFactoryTest;
 import io.grpc.internal.ClientTransportFactory;
-import io.grpc.internal.testing.AbstractClientTransportFactoryTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
